fix(NgRedux): Make App Ref Optional Dep
When we changed to how we configuring a store in V3, it started to be
done in the root component of the application that needed an
ApplicationRef to be injected.

ApplicationRef is primarly needed for the Redux Dev Tools, which we
don't need during testing.

Changed this to an optional dependency, and a check to see if it's
defined before calling tick.

When setting up a mock of NgRedux to be used in unit tests, you need to
provide a configured store.

angular2-redux-starter has [pull
request](https://github.com/rangle/angular2-redux-starter/pull/186)
showing this, but the basics are:

```js
beforeEachProviders(() => [RioCounterPage, CounterActions,
provide(NgRedux, {
    useFactory: () => {
      let ngRedux = new NgRedux(null);
      ngRedux.configureStore(combineReducers<any>({ counter }), {});
      return ngRedux;
    }
  })]);
```

This will provide a configured instance of NgRedux to be injected into
components and not throw an error during unit tests.